+++ apecs-box2d/demo/Main.hs
@@ -94,4 +94,4 @@ main = do
   w <- initWorld
   runSystem (initialize >> play disp black 60 draw handle step) w
   where
-    disp = InWindow "apecs-box2d demo" (720, 720) (10, 10)
+    disp = InWindow "apecs-box2d tumbler" (720, 720) (10, 10)